Integrate Feature Policy: "sync-xhr"
This adds a policy-controlled feature, named 'sync-xhr', which can be disabled in a document to turn off synchronous requests for that document (and documents in all descendant frames). Calling send() on a synchronous request in a document where "sync-xhr" is disabled will result in a "NetworkError" DOMException exception being thrown.

Caveat: https://github.com/whatwg/html/pull/3287 which redefines "allowed to use" in HTML to be more like https://wicg.github.io/feature-policy/#allowed-to-use has not yet landed. If that takes significant time we should add a note to its usage here.

Tests: xhr/xmlhttprequest-sync-default-feature-policy.sub.html in web-platform-tests.

Fixes #178.