Specify target of progress events
The algorithm for "fire a progress event" requires a name and a target,
but only half of the 14 usages specify a target. This is ambiguous
because both the `XMLHttpRequestEventTarget` and the
`XMLHttpRequestUpload` support a "progress" event.

Consistently specify the intended target with each invocation. Also
modify a preposition in the algorithm's definition to match the
prevailing usage.